New logo + other stuff

+++ Source/HedgeModManager.UI/CLI/Commands/CliCommandContinueUpdate.cs
@@ -30,24 +30,34 @@ public class CliCommandContinueUpdate : ICliCommand
         switch (Stage)
         {
             case 0: // Copy files
+
+                bool isSingleFile = File.Exists(Path.Combine(Program.InstallLocation, "update.exe"));
                 var updateDir = Program.InstallLocation;
 
-                if (Directory.Exists(TargetDir))
-                    Directory.Delete(TargetDir, true);
-                foreach (var subDir in Directory.GetDirectories(updateDir, "*", SearchOption.AllDirectories))
-                {
-                    var targetSubDir = TargetDir + subDir[updateDir.Length..];
-                    Directory.CreateDirectory(targetSubDir);
-                    foreach (var file in Directory.GetFiles(subDir))
-                        File.Copy(file, Path.Combine(targetSubDir, Path.GetFileName(file)));
-                }
-                // Root files
-                foreach (var file in Directory.GetFiles(updateDir))
-                    File.Copy(file, Path.Combine(TargetDir, Path.GetFileName(file)));
-
-                string exePath = Path.Combine(TargetDir, "HedgeModManager.UI.exe");
+                string exePath = Path.Combine(TargetDir, "HedgeModManager.exe");
                 if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
                     exePath = Path.Combine(TargetDir, "HedgeModManager.UI");
+
+                if (isSingleFile)
+                {
+                    string packagePath = Path.Combine(updateDir, "update.exe");
+                    File.Copy(packagePath, exePath, true);
+                }
+                else
+                {
+                    if (Directory.Exists(TargetDir))
+                        Directory.Delete(TargetDir, true);
+                    foreach (var subDir in Directory.GetDirectories(updateDir, "*", SearchOption.AllDirectories))
+                    {
+                        var targetSubDir = TargetDir + subDir[updateDir.Length..];
+                        Directory.CreateDirectory(targetSubDir);
+                        foreach (var file in Directory.GetFiles(subDir))
+                            File.Copy(file, Path.Combine(targetSubDir, Path.GetFileName(file)));
+                    }
+                    // Root files
+                    foreach (var file in Directory.GetFiles(updateDir))
+                        File.Copy(file, Path.Combine(TargetDir, Path.GetFileName(file)));
+                }
                 string args = "--continue-update";
                 args += $" {Process.GetCurrentProcess().Id}";
                 args += $" 1";
